fix: Fix INTEGER overflow for large energy values and duplicate reference numbers
- Migrate consumption_previous_year, consumption_forecast, feed_in_forecast from INTEGER to BIGINT
  to support values exceeding 2.1B (INT4 max), observed in production with values like 12222222222
- Change corresponding Go types from *int to *int64 across models, requests, repo scan, service
- Replace generateReferenceNumber() Unix-timestamp suffix with crypto/rand (range 100000-999999)
  to eliminate duplicate reference number constraint violations under concurrent load

+++ internal/shared/models.go
@@ -105,9 +105,9 @@ type Application struct {
 	// Configurable application-level fields (PROJ-8)
 	MembershipStartDate     *time.Time `json:"membershipStartDate,omitempty" db:"membership_start_date"`
 	PersonsInHousehold      *int       `json:"personsInHousehold,omitempty" db:"persons_in_household"`
-	ConsumptionPreviousYear *int       `json:"consumptionPreviousYear,omitempty" db:"consumption_previous_year"`
-	ConsumptionForecast     *int       `json:"consumptionForecast,omitempty" db:"consumption_forecast"`
-	FeedInForecast          *int       `json:"feedInForecast,omitempty" db:"feed_in_forecast"`
+	ConsumptionPreviousYear *int64     `json:"consumptionPreviousYear,omitempty" db:"consumption_previous_year"`
+	ConsumptionForecast     *int64     `json:"consumptionForecast,omitempty" db:"consumption_forecast"`
+	FeedInForecast          *int64     `json:"feedInForecast,omitempty" db:"feed_in_forecast"`
 	PvPowerKwp              *float64   `json:"pvPowerKwp,omitempty" db:"pv_power_kwp"`
 	HeatPump                *bool      `json:"heatPump,omitempty" db:"heat_pump"`
 	ElectricVehicle         *bool      `json:"electricVehicle,omitempty" db:"electric_vehicle"`
